decompose flamegraph code/fix minor bug
