+++ Profiler_panel/devtools.html
@@ -1,4 +1,3 @@
-<!-- devtools.html -->
 <!doctype html>
 <html>
   <head>
