+++ Profiler_panel/popup.html
@@ -37,7 +37,6 @@
         <div>
           <label for="extensionDropdown">Extension IDs:</label>
           <select id="extensionDropdown">
-            <!-- Options will be dynamically added here -->
           </select>
         </div>
         <div class="link-container">
